Get a data frame containing the available SBSs

This method returns a data frame containing the available SBSs and the corresponding mutation probability. The first column ("Type") describes a mutation in a context, while each of the remaining columns contains the probabilities of the mutations for one of the available SBSs.

Value

A data frame containing the available SBSs.

Examples

# build a mutation engine
m_engine <- build_mutation_engine(setup_code = "demo")

# loading dplyr to filter `get_SBSs()` output
library("dplyr")
#> 
#> Attaching package: ‘dplyr’
#> The following objects are masked from ‘package:stats’:
#> 
#>     filter, lag
#> The following objects are masked from ‘package:base’:
#> 
#>     intersect, setdiff, setequal, union

# get the SBS data frame
m_engine$get_SBSs() %>% head()
#>      Type         SBS1         SBS2        SBS3        SBS4        SBS5
#> 1 A[C>A]A 0.0008760226 5.789892e-07 0.020919749 0.042450737 0.012052048
#> 2 A[C>A]C 0.0022201200 1.455043e-04 0.016343192 0.032989582 0.009337087
#> 3 A[C>A]G 0.0001797268 5.361706e-05 0.001807548 0.016115961 0.001908140
#> 4 A[C>A]T 0.0012650525 9.758841e-05 0.012265066 0.029662770 0.006635871
#> 5 A[C>G]A 0.0018390547 2.226114e-16 0.019813416 0.006930938 0.010143807
#> 6 A[C>G]C 0.0011879592 1.307576e-04 0.011588809 0.002813526 0.005637860
#>           SBS6        SBS7a        SBS7b        SBS7c        SBS7d        SBS8
#> 1 0.0004251456 6.712587e-05 2.344280e-03 0.0048409877 4.043656e-05 0.044310640
#> 2 0.0005162257 1.767296e-04 4.567882e-04 0.0011351274 7.539588e-04 0.047299564
#> 3 0.0000533916 7.325781e-05 1.921876e-04 0.0003880497 2.569364e-04 0.004767275
#> 4 0.0001799861 2.485224e-04 7.140506e-04 0.0019636310 4.051334e-03 0.047204587
#> 5 0.0004711612 6.502192e-05 8.614734e-06 0.0011225477 1.180887e-03 0.004350682
#> 6 0.0002847120 2.131222e-05 1.872728e-04 0.0003148746 2.513195e-03 0.002919116
#>           SBS9       SBS10a       SBS10b       SBS10c       SBS10d        SBS11
#> 1 0.0005606393 2.185262e-03 1.809142e-04 0.0043426398 0.0101247301 1.473375e-04
#> 2 0.0040469910 1.739372e-03 6.402356e-03 0.0146429168 0.0181854988 5.486041e-04
#> 3 0.0004395597 1.537130e-04 5.461614e-05 0.0006765830 0.0007469252 9.748818e-05
#> 4 0.0030631330 1.695607e-02 1.619619e-05 0.0131561203 0.0142057670 2.683238e-04
#> 5 0.0048628896 2.225003e-16 3.031814e-04 0.0003485318 0.0001290341 1.715572e-04
#> 6 0.0019393888 2.387766e-05 1.311797e-04 0.0003210247 0.0002879699 2.882160e-04
#>          SBS12        SBS13        SBS14        SBS15       SBS16       SBS17a
#> 1 0.0045423679 0.0018168793 1.126093e-03 9.460299e-04 0.016094373 2.072799e-03
#> 2 0.0011183613 0.0007088422 1.297141e-02 4.905110e-04 0.002872842 9.052935e-04
#> 3 0.0005573098 0.0002706555 4.264466e-04 4.743509e-05 0.001053691 4.894382e-05
#> 4 0.0012255208 0.0003472572 8.301435e-02 1.111919e-03 0.010658028 6.184968e-05
#> 5 0.0012260376 0.0038633644 2.231097e-16 1.132428e-04 0.001750263 1.011366e-03
#> 6 0.0006353875 0.0009349641 5.720752e-06 2.654885e-04 0.001991177 5.611243e-04
#>         SBS17b       SBS18        SBS19        SBS20        SBS21       SBS22a
#> 1 6.077685e-04 0.051688322 0.0012796133 6.242471e-04 1.577225e-04 6.079509e-03
#> 2 1.269942e-04 0.015617160 0.0006360522 1.380513e-03 2.334882e-03 9.357927e-05
#> 3 5.975117e-05 0.002504669 0.0002545473 2.256746e-05 3.033185e-04 8.161555e-04
#> 4 4.556347e-04 0.021469202 0.0005750795 1.249985e-03 6.235937e-04 1.738427e-03
#> 5 1.459445e-04 0.001736326 0.0032141460 8.844349e-03 1.697853e-05 1.679199e-03
#> 6 4.272815e-05 0.002560028 0.0017761836 1.499695e-03 7.113472e-04 1.913146e-04
#>        SBS22b        SBS23       SBS24       SBS25        SBS26       SBS27
#> 1 0.006129164 8.445069e-04 0.036623727 0.009935785 0.0008768747 0.005244790
#> 2 0.003771034 3.974198e-04 0.026258314 0.006918705 0.0005222960 0.004699250
#> 3 0.001013499 1.023483e-07 0.015292613 0.001494492 0.0001175941 0.000809489
#> 4 0.003589467 5.667931e-18 0.022125890 0.004985863 0.0006214856 0.002737009
#> 5 0.004208605 5.801719e-18 0.003360529 0.008059026 0.0004288955 0.001328814
#> 6 0.002689769 4.061297e-05 0.003636529 0.001620954 0.0003739166 0.001754783
#>          SBS28       SBS29        SBS30       SBS31       SBS32        SBS33
#> 1 0.0007856708 0.063497599 1.810502e-03 0.009609122 0.022378921 0.0031154137
#> 2 0.0024969219 0.050660560 5.012295e-04 0.018351301 0.018580266 0.0021999862
#> 3 0.0003632936 0.019810722 9.432639e-05 0.001717126 0.004596507 0.0004259067
#> 4 0.0039767846 0.035752559 5.590079e-04 0.006322835 0.014645495 0.0019025053
#> 5 0.0001543278 0.003154786 1.076243e-03 0.008380282 0.001997043 0.0014124538
#> 6 0.0002210714 0.002602290 4.625973e-04 0.003134601 0.002006273 0.0012825028
#>          SBS34       SBS35       SBS36       SBS37       SBS38       SBS39
#> 1 4.914517e-03 0.008898800 0.025256687 0.003963390 0.012902366 0.011747931
#> 2 6.917052e-03 0.045853499 0.008212189 0.001432840 0.010026298 0.007070358
#> 3 5.391344e-05 0.001438615 0.002305593 0.001092282 0.001966852 0.002753250
#> 4 1.240720e-03 0.021759121 0.017932705 0.001855489 0.008917017 0.007427184
#> 5 7.689654e-04 0.003607894 0.001844139 0.034416274 0.002711513 0.047493774
#> 6 1.123027e-03 0.004515876 0.001648361 0.011067450 0.001568471 0.022249377
#>        SBS40a      SBS40b      SBS40c        SBS41        SBS42        SBS43
#> 1 0.036537452 0.014058734 0.003775855 2.119925e-03 1.170674e-03 2.958436e-02
#> 2 0.016582512 0.004665315 0.006217657 1.207140e-03 2.047411e-02 5.827094e-03
#> 3 0.003864093 0.001174707 0.002083570 6.334691e-05 3.460899e-05 6.447238e-04
#> 4 0.015488554 0.004969113 0.008211689 1.335693e-03 8.039950e-03 2.075045e-03
#> 5 0.008245511 0.007480618 0.008936205 5.355071e-03 4.097357e-04 3.620324e-04
#> 6 0.005556710 0.010095786 0.004277028 3.126690e-03 5.029078e-04 3.090663e-05
#>          SBS44       SBS45        SBS46        SBS47        SBS48        SBS49
#> 1 7.723217e-18 0.009177748 0.0044121196 0.0681598308 8.464570e-04 0.0252196531
#> 2 1.485929e-04 0.002827629 0.0046414349 0.0294045851 8.199633e-04 0.0069563358
#> 3 9.460004e-07 0.001717452 0.0003089404 0.0014247436 2.623113e-04 0.0073365833
#> 4 5.811548e-03 0.009707595 0.0034079281 0.0119581138 3.847403e-05 0.0080046186
#> 5 3.198705e-03 0.003133128 0.0012033054 0.0010354667 1.960213e-04 0.0002893731
#> 6 2.308143e-04 0.001914850 0.0014813090 0.0006108633 1.111481e-04 0.0001335849
#>         SBS50       SBS51       SBS52        SBS53        SBS54        SBS55
#> 1 0.119346280 0.141438096 0.015188294 5.414513e-03 0.0021700061 5.899034e-03
#> 2 0.125436939 0.001669528 0.006435806 1.942643e-03 0.0007875536 2.025430e-03
#> 3 0.005942874 0.004944785 0.004248392 3.937866e-02 0.0016920363 4.625147e-05
#> 4 0.083406934 0.021256906 0.009228992 1.719123e-05 0.0004157429 1.173292e-03
#> 5 0.003861203 0.003972303 0.001728667 7.075098e-03 0.0015571804 2.859225e-03
#> 6 0.001086462 0.004485005 0.001495784 5.411648e-03 0.0004907367 1.659863e-04
#>          SBS56       SBS57        SBS58        SBS59        SBS60        SBS84
#> 1 0.0126059900 0.012329125 0.0589648977 0.0036167178 6.163154e-03 0.0035002399
#> 2 0.0154691290 0.001411637 0.0066549043 0.0023514098 7.688209e-04 0.0049714481
#> 3 0.0002116571 0.003427917 0.0008481854 0.0001469153 2.294037e-16 0.0004686325
#> 4 0.0230012344 0.008656808 0.0055837933 0.0146017560 4.447616e-04 0.0093676200
#> 5 0.0004181983 0.015436467 0.0206226979 0.0029618809 3.778064e-04 0.0068130534
#> 6 0.0001635591 0.003089807 0.0017154863 0.0024208607 5.803167e-04 0.0107345240
#>          SBS85        SBS86       SBS87        SBS88       SBS89        SBS90
#> 1 0.0061078673 0.0029681674 0.008946076 1.004268e-18 0.032296782 0.0022220019
#> 2 0.0008705206 0.0037346766 0.004489906 9.890296e-19 0.017494551 0.0007039243
#> 3 0.0003156920 0.0003977458 0.006357213 1.031355e-18 0.009971177 0.0001436189
#> 4 0.0027284676 0.0036394761 0.004940762 1.737757e-03 0.020817534 0.0017706390
#> 5 0.0072677806 0.0527625043 0.007843246 1.004268e-18 0.014876093 0.0005127696
#> 6 0.0025240523 0.0774946614 0.003862856 3.994926e-03 0.005816986 0.0002754479
#>          SBS91       SBS92       SBS93       SBS94       SBS95       SBS96
#> 1 0.0029343969 0.011395596 0.011627705 0.015676823 0.038614422 0.011463170
#> 2 0.0520129749 0.009652994 0.008011028 0.024522707 0.017211907 0.007674629
#> 3 0.0002089847 0.004851456 0.001816620 0.001626815 0.008631699 0.002056911
#> 4 0.0001303040 0.007800061 0.008456891 0.011140517 0.023409398 0.011837910
#> 5 0.0002424800 0.003073731 0.008898488 0.007047547 0.003636111 0.009321073
#> 6 0.0001222608 0.001673952 0.005361007 0.003954052 0.002778859 0.002500765
#>         SBS97       SBS98        SBS99
#> 1 0.008222233 0.008775961 3.853267e-06
#> 2 0.007456078 0.006644665 6.914087e-06
#> 3 0.001422664 0.033739961 5.324609e-05
#> 4 0.006908484 0.003744396 3.095431e-05
#> 5 0.002365081 0.006874813 4.499113e-04
#> 6 0.002204869 0.003205283 1.209278e-02