Getting the active germline subject

This method returns the active germline subject.

Value

A dataframe the active germline subject.

Details

The active germline subject is returned as a dataframe in which the column sample reports the subject name, the columns pop and super_pop contain the subject population and super population, respectively, and the column gender declares the subject gender.

See also

Examples

# build a mutation engine
m_engine <- MutationEngine(setup_code = "demo")
#>  [█---------------------------------------] 0% [00m:00s] Loading context index                                 [████████████████████████████████████████] 100% [00m:00s] Context index loaded                               
#>  [█---------------------------------------] 0% [00m:00s] Loading RS index                                      [██████████------------------------------] 23% [00m:01s] Loading RS index                                     [████████████████████--------------------] 49% [00m:02s] Loading RS index                                     [█████████████████████████████-----------] 71% [00m:03s] Loading RS index                                     [██████████████████████████████████████--] 94% [00m:04s] Loading RS index                                     [████████████████████████████████████████] 100% [00m:04s] RS index loaded                                    
#>  [█---------------------------------------] 0% [00m:00s] Loading germline                                      [████████████████████████████████████████] 100% [00m:00s] Germline loaded                                    

# get the active germline subject dataframe
head(m_engine$get_active_germline(), 5)
#>    sample pop super_pop gender
#> 1 NA18941 JPT       EAS female