Getting the known driver mutations

This method returns the known driver mutations.

Value

A dataframe containing the known driver.

Details

The mutation are returned in a dataframe reporting the known driver mutations together with their types, associated tumours, affected genes, and code name. The first three columns ("chr", "from", and "to") report the mutation chromosome, the initial position and the final position, respectively. The next three columns ("ref", "alt", and "mutation_type") describe the reference sequence, the altered sequence, and the type of the mutation. The last three columns ("tumour_type", "driver_gene", and "driver_code") detail the tumour type associated to the mutation, the affected gene, and the driver code, which can be used to specify the mutation when adding mutants to the mutation engine.

See also

Examples

# build a mutation engine
m_engine <- MutationEngine(setup_code = "demo")
#>  [█---------------------------------------] 0% [00m:00s] Loading context index                                 [████████████████████████████████████████] 100% [00m:00s] Context index loaded                               
#>  [█---------------------------------------] 0% [00m:00s] Loading RS index                                      [█████████-------------------------------] 22% [00m:01s] Loading RS index                                     [████████████████------------------------] 39% [00m:02s] Loading RS index                                     [█████████████████████████---------------] 62% [00m:03s] Loading RS index                                     [██████████████████████████████████------] 84% [00m:04s] Loading RS index                                     [████████████████████████████████████████] 100% [00m:04s] RS index loaded                                    
#>  [█---------------------------------------] 0% [00m:00s] Loading germline                                      [████████████████████████████████████████] 100% [00m:00s] Germline loaded                                    

# get the known driver dataframe
head(m_engine$get_known_drivers(), 5)
#>   chr      from        to     ref alt mutation_type tumour_type driver_gene
#> 1   1 115256502 115256508 TGGTCTC   -         indel        PACA        NRAS
#> 2   1 115256502 115256508 TGGTCTC   -         indel        PACA        NRAS
#> 3   1 115256528 115256528       T   G           SNV        MALY        NRAS
#> 4   1 115256529 115256529       T   C           SNV        MELA        NRAS
#> 5   1 115256528 115256528       T   G           SNV        MALY        NRAS
#>   driver_code
#> 1  NRAS RDQ68
#> 2    NRAS A66
#> 3   NRAS Q61H
#> 4   NRAS Q61R
#> 5   NRAS Q61K