Retrieving samples' information

This method retrieves information about the samples whose cells were used as leaves of the samples forest.

Value

A dataframe containing, for each sample collected during the simulation, the columns "name", "time", "id", "ymin", "xmin", "ymax", "xmax", "tumour_cells", and "tumour_cells_in_bbox". The columns "ymin", "xmin", "ymax", "xmax" report the boundaries of the sample bounding box, while "tumour_cells" and "tumour_cells_in_bbox" are the number of tumour cells in the sample and in the bounding box, respectively.

See also

Examples

# set the seed of the random number generator
set.seed(0)

# create a simulation
sim <- SpatialSimulation()
sim$add_mutant(name = "A", growth_rate = 0.2,
               death_rate = 0.01)
sim$place_cell("A", 500, 500)

sim$death_activation_level <- 100
sim$run_up_to_size(species = "A", num_of_cells = 50000)
#>  [██████████████████████████████████------] 84% [00m:00s] Cells: 42285                                         [████████████████████████████████████████] 100% [00m:00s] Saving snapshot                                    

# sample the region [450,500]x[475,550]
sim$sample_cells("S1", lower_corner=c(450,475), upper_corner=c(500,550))

# build the samples forest
forest <- sim$get_samples_forest()

# get information about the sampled whose cells
# are the forest leaves, i.e, S1 and S2
forest$get_samples_info()
#>   name id xmin ymin xmax ymax tumour_cells tumour_cells_in_bbox     time
#> 1   S1  5  450  475  500  550         3863                 3863 215.0761