Gets the tissue cells

This method collects some data about the cells in the tissue without altering the tissue itself. The pairs of optional parameters lower_corner and upper_corner define a frame of the tissue in which the data are sampled. The optional parameters mutant_filter and epigenetic_filter filter the collected cell data according to the cell mutant and epigenetic state.

Arguments

lower_corner

The lower-left corner of the selection frame (optional).

upper_corner

The upper-right corner of the selection frame (optional).

mutant_filter

The vector of the to-be-selected mutant names (optional).

epigenetic_filter

The vector of the to-be-selected epigenetic states (optional).

Value

A data frame reporting cell_id, mutant, epistate, position_x, and position_y for each cells satisfying the provided filters and laying in the input frame.

Examples

sim <- new(Simulation)
sim$add_mutant(name = "A",
               epigenetic_rates = c("+-" = 0.01, "-+" = 0.01),
               growth_rates = c("+" = 0.2, "-" = 0.08),
               death_rates = c("+" = 0.1, "-" = 0.01))
sim$add_mutant(name = "B",
               epigenetic_rates = c("+-" = 0.02, "-+" = 0.01),
               growth_rates = c("+" = 0.3, "-" = 0.1),
               death_rates = c("+" = 0.1, "-" = 0.01))
sim$schedule_mutation(src = "A", dst = "B", time = 50)
sim$place_cell("A+", 500, 500)
sim$run_up_to_time(30)
#>  [████████████████████████████████████████] 100% [00m:00s] Saving snapshot                                        

# collect all the cells in the tissue
sim$get_cells()
#>    cell_id mutant epistate position_x position_y
#> 1      250      A        -        495        501
#> 2      249      A        +        496        500
#> 3      215      A        +        497        501
#> 4      216      A        +        497        502
#> 5      161      A        -        498        499
#> 6       82      A        -        498        502
#> 7      177      A        +        498        504
#> 8      199      A        +        499        497
#> 9      190      A        +        499        499
#> 10     225      A        +        499        500
#> 11     162      A        -        499        501
#> 12     226      A        +        499        502
#> 13     151      A        +        499        503
#> 14     204      A        +        499        504
#> 15     230      A        +        499        505
#> 16     214      A        +        500        495
#> 17     212      A        +        500        498
#> 18     189      A        +        500        499
#> 19     127      A        -        500        500
#> 20     242      A        +        500        501
#> 21     246      A        +        500        502
#> 22     203      A        +        500        503
#> 23     233      A        +        500        504
#> 24     229      A        +        500        505
#> 25     237      A        +        500        507
#> 26     232      A        +        500        508
#> 27     207      A        +        501        496
#> 28     239      A        +        501        497
#> 29     200      A        +        501        498
#> 30     211      A        +        501        499
#> 31     179      A        +        501        500
#> 32     192      A        +        501        501
#> 33     241      A        +        501        502
#> 34     245      A        +        501        503
#> 35     174      A        +        501        504
#> 36     234      A        +        501        505
#> 37     169      A        +        501        506
#> 38     238      A        +        501        507
#> 39     240      A        +        502        496
#> 40     171      A        +        502        497
#> 41     213      A        -        502        498
#> 42     198      A        +        502        500
#> 43      81      A        -        502        501
#> 44     168      A        +        502        502
#> 45     236      A        +        502        503
#> 46     235      A        +        502        504
#> 47     175      A        +        502        505
#> 48     220      A        +        502        506
#> 49     206      A        +        502        507
#> 50     195      A        +        503        496
#> 51     254      A        +        503        498
#> 52     197      A        +        503        499
#> 53     191      A        +        503        500
#> 54     180      A        +        503        501
#> 55     186      A        +        503        502
#> 56     176      A        +        503        503
#> 57     222      A        +        503        504
#> 58     221      A        +        503        505
#> 59     205      A        +        503        507
#> 60     208      A        +        504        496
#> 61     224      A        +        504        497
#> 62     228      A        +        504        498
#> 63     253      A        +        504        499
#> 64     251      A        -        504        500
#> 65     247      A        +        504        501
#> 66     184      A        +        504        502
#> 67     182      A        +        504        503
#> 68     218      A        +        504        504
#> 69     217      A        +        504        505
#> 70     158      A        +        504        506
#> 71     219      A        +        504        507
#> 72     196      A        +        505        497
#> 73     223      A        +        505        498
#> 74     227      A        +        505        499
#> 75     252      A        -        505        500
#> 76     201      A        +        505        501
#> 77     248      A        +        505        502
#> 78     209      A        +        505        503
#> 79     187      A        +        505        505
#> 80     188      A        +        505        506
#> 81     244      A        +        506        500
#> 82     243      A        +        506        501
#> 83     181      A        +        506        503
#> 84     166      A        +        506        504
#> 85     210      A        +        507        504

# get the cells in the frame [495,505]x[490,500]
sim$get_cells(lower_corner=c(495,490), upper_corner=c(505,500))
#>    cell_id mutant epistate position_x position_y
#> 1      249      A        +        496        500
#> 2      161      A        -        498        499
#> 3      199      A        +        499        497
#> 4      190      A        +        499        499
#> 5      225      A        +        499        500
#> 6      214      A        +        500        495
#> 7      212      A        +        500        498
#> 8      189      A        +        500        499
#> 9      127      A        -        500        500
#> 10     207      A        +        501        496
#> 11     239      A        +        501        497
#> 12     200      A        +        501        498
#> 13     211      A        +        501        499
#> 14     179      A        +        501        500
#> 15     240      A        +        502        496
#> 16     171      A        +        502        497
#> 17     213      A        -        502        498
#> 18     198      A        +        502        500
#> 19     195      A        +        503        496
#> 20     254      A        +        503        498
#> 21     197      A        +        503        499
#> 22     191      A        +        503        500
#> 23     208      A        +        504        496
#> 24     224      A        +        504        497
#> 25     228      A        +        504        498
#> 26     253      A        +        504        499
#> 27     251      A        -        504        500
#> 28     196      A        +        505        497
#> 29     223      A        +        505        498
#> 30     227      A        +        505        499
#> 31     252      A        -        505        500

# cells can be filtered by mutant name...
sim$get_cells(mutant_filter=c("A"),epigenetic_filter=c("+","-"))
#>    cell_id mutant epistate position_x position_y
#> 1      250      A        -        495        501
#> 2      249      A        +        496        500
#> 3      215      A        +        497        501
#> 4      216      A        +        497        502
#> 5      161      A        -        498        499
#> 6       82      A        -        498        502
#> 7      177      A        +        498        504
#> 8      199      A        +        499        497
#> 9      190      A        +        499        499
#> 10     225      A        +        499        500
#> 11     162      A        -        499        501
#> 12     226      A        +        499        502
#> 13     151      A        +        499        503
#> 14     204      A        +        499        504
#> 15     230      A        +        499        505
#> 16     214      A        +        500        495
#> 17     212      A        +        500        498
#> 18     189      A        +        500        499
#> 19     127      A        -        500        500
#> 20     242      A        +        500        501
#> 21     246      A        +        500        502
#> 22     203      A        +        500        503
#> 23     233      A        +        500        504
#> 24     229      A        +        500        505
#> 25     237      A        +        500        507
#> 26     232      A        +        500        508
#> 27     207      A        +        501        496
#> 28     239      A        +        501        497
#> 29     200      A        +        501        498
#> 30     211      A        +        501        499
#> 31     179      A        +        501        500
#> 32     192      A        +        501        501
#> 33     241      A        +        501        502
#> 34     245      A        +        501        503
#> 35     174      A        +        501        504
#> 36     234      A        +        501        505
#> 37     169      A        +        501        506
#> 38     238      A        +        501        507
#> 39     240      A        +        502        496
#> 40     171      A        +        502        497
#> 41     213      A        -        502        498
#> 42     198      A        +        502        500
#> 43      81      A        -        502        501
#> 44     168      A        +        502        502
#> 45     236      A        +        502        503
#> 46     235      A        +        502        504
#> 47     175      A        +        502        505
#> 48     220      A        +        502        506
#> 49     206      A        +        502        507
#> 50     195      A        +        503        496
#> 51     254      A        +        503        498
#> 52     197      A        +        503        499
#> 53     191      A        +        503        500
#> 54     180      A        +        503        501
#> 55     186      A        +        503        502
#> 56     176      A        +        503        503
#> 57     222      A        +        503        504
#> 58     221      A        +        503        505
#> 59     205      A        +        503        507
#> 60     208      A        +        504        496
#> 61     224      A        +        504        497
#> 62     228      A        +        504        498
#> 63     253      A        +        504        499
#> 64     251      A        -        504        500
#> 65     247      A        +        504        501
#> 66     184      A        +        504        502
#> 67     182      A        +        504        503
#> 68     218      A        +        504        504
#> 69     217      A        +        504        505
#> 70     158      A        +        504        506
#> 71     219      A        +        504        507
#> 72     196      A        +        505        497
#> 73     223      A        +        505        498
#> 74     227      A        +        505        499
#> 75     252      A        -        505        500
#> 76     201      A        +        505        501
#> 77     248      A        +        505        502
#> 78     209      A        +        505        503
#> 79     187      A        +        505        505
#> 80     188      A        +        505        506
#> 81     244      A        +        506        500
#> 82     243      A        +        506        501
#> 83     181      A        +        506        503
#> 84     166      A        +        506        504
#> 85     210      A        +        507        504

# ...or by epigenetic state
sim$get_cells(mutant_filter=c("A","B"),epigenetic_filter=c("-"))
#>   cell_id mutant epistate position_x position_y
#> 1     250      A        -        495        501
#> 2     161      A        -        498        499
#> 3      82      A        -        498        502
#> 4     162      A        -        499        501
#> 5     127      A        -        500        500
#> 6     213      A        -        502        498
#> 7      81      A        -        502        501
#> 8     251      A        -        504        500
#> 9     252      A        -        505        500

# cells can be filtered by frame, mutant, and epigenetic states
sim$get_cells(lower_corner=c(495,495), upper_corner=c(505,505),
              mutant_filter=c("A"),epigenetic_filter=c("+","-"))
#>    cell_id mutant epistate position_x position_y
#> 1      250      A        -        495        501
#> 2      249      A        +        496        500
#> 3      215      A        +        497        501
#> 4      216      A        +        497        502
#> 5      161      A        -        498        499
#> 6       82      A        -        498        502
#> 7      177      A        +        498        504
#> 8      199      A        +        499        497
#> 9      190      A        +        499        499
#> 10     225      A        +        499        500
#> 11     162      A        -        499        501
#> 12     226      A        +        499        502
#> 13     151      A        +        499        503
#> 14     204      A        +        499        504
#> 15     230      A        +        499        505
#> 16     214      A        +        500        495
#> 17     212      A        +        500        498
#> 18     189      A        +        500        499
#> 19     127      A        -        500        500
#> 20     242      A        +        500        501
#> 21     246      A        +        500        502
#> 22     203      A        +        500        503
#> 23     233      A        +        500        504
#> 24     229      A        +        500        505
#> 25     207      A        +        501        496
#> 26     239      A        +        501        497
#> 27     200      A        +        501        498
#> 28     211      A        +        501        499
#> 29     179      A        +        501        500
#> 30     192      A        +        501        501
#> 31     241      A        +        501        502
#> 32     245      A        +        501        503
#> 33     174      A        +        501        504
#> 34     234      A        +        501        505
#> 35     240      A        +        502        496
#> 36     171      A        +        502        497
#> 37     213      A        -        502        498
#> 38     198      A        +        502        500
#> 39      81      A        -        502        501
#> 40     168      A        +        502        502
#> 41     236      A        +        502        503
#> 42     235      A        +        502        504
#> 43     175      A        +        502        505
#> 44     195      A        +        503        496
#> 45     254      A        +        503        498
#> 46     197      A        +        503        499
#> 47     191      A        +        503        500
#> 48     180      A        +        503        501
#> 49     186      A        +        503        502
#> 50     176      A        +        503        503
#> 51     222      A        +        503        504
#> 52     221      A        +        503        505
#> 53     208      A        +        504        496
#> 54     224      A        +        504        497
#> 55     228      A        +        504        498
#> 56     253      A        +        504        499
#> 57     251      A        -        504        500
#> 58     247      A        +        504        501
#> 59     184      A        +        504        502
#> 60     182      A        +        504        503
#> 61     218      A        +        504        504
#> 62     217      A        +        504        505
#> 63     196      A        +        505        497
#> 64     223      A        +        505        498
#> 65     227      A        +        505        499
#> 66     252      A        -        505        500
#> 67     201      A        +        505        501
#> 68     248      A        +        505        502
#> 69     209      A        +        505        503
#> 70     187      A        +        505        505