Gets the history of the number of cells per species

This method returns a data frame reporting the number of species cells in each sampled simulation time.

Value

A data frame reporting "mutant", "epistate", "counts", and "time" for each species, and for each sampled time.

Examples

sim <- new(Simulation)
sim$add_mutant("A", growth_rate = 0.2, death_rate = 0.1)
sim$add_mutant("B", growth_rate = 0.15, death_rate = 0.05)
sim$schedule_mutation(src = "A", dst = "B", time = 50)
sim$place_cell("A", 500, 500)
sim$history_delta <- 20
sim$run_up_to_time(70)

# get the history of species counts
sim$get_count_history()
#>   mutant epistate count     time
#> 1      A              4 20.00000
#> 2      B              0 20.00000
#> 3      A             13 40.00000
#> 4      B              0 40.00000
#> 5      A             59 60.00000
#> 6      B              4 60.00000
#> 7      A            136 70.00615
#> 8      B              4 70.00615