Constructs a new Simulation

Constructs a new Simulation

Arguments

simulation_name

The name of the simulation (optional).

seed

The seed for the pseudo-random generator (optional).

save_snapshots

A flag to save simulation snapshots on disk (optional, default `FALSE`).

Examples

# create a Simulation object storing binary dump in a temporary directory. 
# The data are deleted from the disk as soon as the object is destroyed.
sim <- new(Simulation, "test")

# add a new species, place a cell in the tissue, and let the simulation evolve.
sim$add_mutant(name = "A", growth_rate = 0.3, death_rate = 0.02)
sim$place_cell("A", 500, 500)
sim$run_up_to_time(30)

# no directory "test" has been created
"test" %in% list.files(".")
#> [1] FALSE

# (let us delete the directory "test" manually)
unlink("test", recursive = TRUE)

# By using the optional parameter `save_snapshots`, we force the 
# simulation to save its progresses in a local directory whose name 
# is the name of the simulation, i.e., "test". This data will be 
# preserved when the simulation object will be destroyed.
sim <- new(Simulation, "test", save_snapshots=TRUE)

# as done above, we add a new species, place a cell in the tissue, and let the 
# simulation evolve.
sim$add_mutant(name = "A", growth_rate = 0.3, death_rate = 0.02)
sim$place_cell("A", 500, 500)
sim$run_up_to_time(30)

# the directory "test" exists and contains a binary dump of 
# sthe simulation.
"test" %in% list.files(".")
#> [1] TRUE

# let us manually delete the "test" directory
unlink("test", recursive = TRUE)

# we can also provide a random seed to the simulation...
sim <- new(Simulation, "test", 13)

# ...or creating a simulation without providing any name. By default, the 
# simulation name will have the following format `races_<date>_<hour>`.
sim <- new(Simulation, 13)