Searching a rectangular tissue sample

This method searches a rectangular tissue sample.

Arguments

min_num_of_cells

A named integer vector reporting the minimum number of cells per species or mutant.

num_of_cells

The number of cells in the searched sample.

width

The width of the searched sample.

height

The height of the searched sample.

Value

If a rectangular sample satisfying the provided constraints can be found, the corresponding rectangle.

Details

The aimed sample mush satisfy the specified number of cells. The sizes of the samples are also provided a parameter of the method. The complexity of this method is \(O(|\textrm{tissue width}|*|\textrm{tissue height}|)\).

See also

Examples

# set the seed of the random number generator
set.seed(0)

# create a simulation
sim <- SpatialSimulation()
sim$death_activation_level <- 50
sim$add_mutant(name = "A", growth_rate = 0.2, death_rate = 0.01)
sim$place_cell("A", 500, 500)
sim$run_up_to_size(species = "A", num_of_cells = 500)
#>  [████████████████████████████████████████] 100% [00m:00s] Saving snapshot                                    

sim$add_mutant(name = "B", growth_rate = 0.3, death_rate = 0.01)
sim$mutate_progeny(sim$choose_cell_in("A"), "B")
sim$run_up_to_size(species = "B", num_of_cells = 1000)
#>  [███-------------------------------------] 6% [00m:00s] Cells: 60842                                          [█████-----------------------------------] 11% [00m:01s] Cells: 96636                                         [███████---------------------------------] 15% [00m:02s] Cells: 125206                                        [███████---------------------------------] 17% [00m:03s] Cells: 142735                                        [█████████-------------------------------] 20% [00m:04s] Cells: 162917                                        [█████████-------------------------------] 22% [00m:05s] Cells: 183952                                        [███████████-----------------------------] 25% [00m:06s] Cells: 204792                                        [████████████----------------------------] 28% [00m:07s] Cells: 223591                                        [████████████----------------------------] 29% [00m:08s] Cells: 241659                                        [██████████████--------------------------] 33% [00m:09s] Cells: 258527                                        [██████████████--------------------------] 34% [00m:10s] Cells: 271905                                        [███████████████-------------------------] 36% [00m:11s] Cells: 284453                                        [████████████████------------------------] 39% [00m:12s] Cells: 299033                                        [█████████████████-----------------------] 41% [00m:13s] Cells: 313570                                        [█████████████████-----------------------] 42% [00m:14s] Cells: 326970                                        [███████████████████---------------------] 46% [00m:15s] Cells: 340360                                        [████████████████████--------------------] 48% [00m:16s] Cells: 353289                                        [████████████████████--------------------] 49% [00m:17s] Cells: 365680                                        [█████████████████████-------------------] 51% [00m:18s] Cells: 377414                                        [██████████████████████------------------] 52% [00m:19s] Cells: 388872                                        [██████████████████████------------------] 54% [00m:20s] Cells: 400553                                        [███████████████████████-----------------] 57% [00m:21s] Cells: 411937                                        [████████████████████████----------------] 58% [00m:22s] Cells: 423194                                        [█████████████████████████---------------] 60% [00m:23s] Cells: 434154                                        [█████████████████████████---------------] 62% [00m:24s] Cells: 445305                                        [██████████████████████████--------------] 63% [00m:25s] Cells: 455883                                        [███████████████████████████-------------] 66% [00m:26s] Cells: 466363                                        [███████████████████████████-------------] 67% [00m:27s] Cells: 476766                                        [█████████████████████████████-----------] 70% [00m:28s] Cells: 486918                                        [█████████████████████████████-----------] 71% [00m:29s] Cells: 497071                                        [██████████████████████████████----------] 74% [00m:30s] Cells: 507122                                        [███████████████████████████████---------] 76% [00m:31s] Cells: 517065                                        [█████████████████████████████████-------] 80% [00m:32s] Cells: 526699                                        [█████████████████████████████████-------] 82% [00m:33s] Cells: 536541                                        [██████████████████████████████████------] 84% [00m:34s] Cells: 546269                                        [███████████████████████████████████-----] 86% [00m:35s] Cells: 555628                                        [████████████████████████████████████----] 88% [00m:36s] Cells: 565034                                        [█████████████████████████████████████---] 90% [00m:37s] Cells: 574263                                        [█████████████████████████████████████---] 91% [00m:38s] Cells: 583547                                        [██████████████████████████████████████--] 94% [00m:39s] Cells: 592710                                        [███████████████████████████████████████-] 95% [00m:40s] Cells: 601707                                        [███████████████████████████████████████-] 96% [00m:41s] Cells: 610602                                        [████████████████████████████████████████] 98% [00m:42s] Cells: 619292                                        [████████████████████████████████████████] 100% [00m:42s] Saving snapshot                                    

# find a 50x50 sample containing 80 "B" cells and 10 "A" cells at least
sim$search_sample(c("A" = 10, "B" = 80), 50, 50)
#> TissueRectangle((358,506),(407,555))