Create a mutation engine object

Create a mutation engine object

Arguments

setup_code

The set-up code (alternative to `directory`).

directory

The set-up directory (alternative to `setup_code`).

reference_url

The reference genome URL (mandatory when `directory` is provided).

SBS_url

The SBS file URL (mandatory when `directory` is provided).

default_num_of_alleles

The default number of alleles per chromosome (mandatory when `directory` is provided).

exceptions_on_allele_number

A list of exceptions for the default number of alleles per chromosome (optional when `directory` in provided).

context_sampling

The number of reference contexts per context in the index (optional: default value is 100).

Value

A mutation engine object.

Details

This function downloads and sets up the data requires by by a mutation engine object and build it.

There are two building modalities: the first one is more general, but it requires to specify all the data sources; the second one adopts some pre-set configurations, but it is more convient than the former in many cases.

The first building modality requires to specify the directory in which the data must be saved, the reference sequence URL, the SBS file URL, the default number of alleles per chromosome, and, optionally, the exceptions on the number of alleles (e.g., one allele for "X" and "Y" in human genome) through the parameters `directory`, `reference_url`, `SBS_url`, `default_num_of_alleles`, and `exceptions_on_allele_number`, respectively.

The second building modality exclusively requires a set-up code (parameter `setup_code`). The list of supported set-up codes can be obtained by using the function `get_mutation_engine_codes`.

The number of context sampling is an optional parameter that allows sampling the reference contexts while building the context index. This parameter, which is set to 100 by default, specifies how many occurences of the same context must be identified before adding one of them to the context index. The larger the number of context sampling, the larger the context index. On the other side, the lower the number of context sampling, the lower the number of sites in the reference genome that can be affected by simulated mutations.

If the parameters of a mutation engine construction match those of a previous construction, then the corresponding reference sequence, the SBS file, and the previously built context index are loaded from the set-up directory avoiding further computations.

See also

Examples

# set the reference and SBS URLs
reference_url <- paste0("https://ftp.ensembl.org/pub/release-110/fasta/",
                        "homo_sapiens/dna/Homo_sapiens.GRCh38.dna.",
                        "chromosome.22.fa.gz")
sbs_url <- paste0("https://cancer.sanger.ac.uk/signatures/documents/2124/",
                  "COSMIC_v3.4_SBS_GRCh38.txt")

# build a mutation engine and save the required files into the
# directory "Test"
m_engine <- build_mutation_engine(directory = "Test",
                                  reference_url = reference_url,
                                  SBS_url = sbs_url,
                                  default_num_of_alleles = 2)
#> Downloading reference genome...
#> Reference genome downloaded
#> Decompressing reference file...done
#> Downloading SBS file...
#> SBS file downloaded
#> Building context index...
#> done

# if the parameters of a mutation engine construction match those of a
# previous construction, then the corresponding reference sequence,
# the SBS file, and the previously built context index are loaded from
# the set-up directory avoiding further computations.
m_engine <- build_mutation_engine("Test", reference_url, sbs_url,
                                  default_num_of_alleles = 2)

# if the `context_sampling` parameter changes, a new context index is
# built, while neither the reference sequence nor the SBS file are 
# downloaded again.
m_engine <- build_mutation_engine("Test", reference_url, sbs_url,
                                  default_num_of_alleles = 2,
                                  context_sampling = 50)
#> Building context index...
#> done

# a futher contruction with the same parameters avoids both downloads
# and context index construction.
m_engine <- build_mutation_engine("Test", reference_url, sbs_url,
                                  default_num_of_alleles = 2,
                                  context_sampling = 50)

# the parameters `directory`, `reference_url`, and `SBS_url` can be
# avoided by providing the `setup_code` parameter. The set-up code 
# `demo` is provided among those available for testing purpose.
m_engine <- build_mutation_engine(setup_code = "demo")

# the `context_sampling` can be used also when a pre-defined set-up
# configuration is adopted.
m_engine <- build_mutation_engine(setup_code = "demo",
                                  context_sampling = 50)
#> Building context index...
#> done

# remove the "Test" and "demo" directories
unlink("Test", recursive = TRUE)
unlink("demo", recursive = TRUE)