Getting the tumour types and studies available for a set-up code.

This method returns the tumour types and studies available for a set-up code.

Arguments

setup_code

The set-up code whose available tumour types and studies are requested.

Value

A dataframe reporting the types and the studies available for a set-up code.

See also

Examples

# get the types and studies available for the "demo" set-up code
get_available_tumours_in("demo")
#>    type study
#> 1  BLCA    US
#> 2  BOCA    UK
#> 3  BRCA    EU
#> 4  BRCA    UK
#> 5  BRCA    US
#> 6  BTCA    SG
#> 7  CESC    US
#> 8  CLLE    ES
#> 9  CMDI    UK
#> 10 COAD    US
#> 11 DLBC    US
#> 12 EOPC    DE
#> 13 ESAD    UK
#> 14 GACA    CN
#> 15  GBM    US
#> 16 HNSC    US
#> 17 KICH    US
#> 18 KIRC    US
#> 19 KIRP    US
#> 20 LAML    KR
#> 21  LGG    US
#> 22 LICA    FR
#> 23 LIHC    US
#> 24 LINC    JP
#> 25 LIRI    JP
#> 26 LUAD    US
#> 27 LUSC    US
#> 28 MALY    DE
#> 29 MELA    AU
#> 30 ORCA    IN
#> 31   OV    AU
#> 32   OV    US
#> 33 PACA    AU
#> 34 PACA    CA
#> 35 PAEN    AU
#> 36 PAEN    IT
#> 37 PBCA    DE
#> 38 PRAD    CA
#> 39 PRAD    UK
#> 40 PRAD    US
#> 41 READ    US
#> 42 RECA    EU
#> 43 SARC    US
#> 44 SKCM    US
#> 45 STAD    US
#> 46 THCA    US
#> 47 UCEC    US