Get mutations-nodes map

Map mutations to nodes of the phylogenetic forest where they occured first.

Usage

muts_to_sticks(phylo_forest, seq_results, labels)

Arguments

phylo_forest

The original forest object where mutations has been placed.

seq_results

Table annotating mutations genome coordinates and causes (the output of simulate_seq()).

labels

Table annotating the sticks (output of get_events_table()).

Value

A tidyverse tibble object

Examples

sim <- new(Simulation)
sim$update_tissue("Liver", 2e3, 2e3)

sim$add_mutant(name = "A",
               growth_rates = 1,
               death_rates = 0)
sim$place_cell("A", 1e3, 1e3)
sim$run_up_to_size("A",1e4)
#>  [████████████████████████████████████████] 100% [00m:00s] Saving snapshot                                                                                                           
sim$add_mutant(name = "B",
               growth_rates = 3.5,
               death_rates = 0)
sim$mutate_progeny(sim$choose_cell_in("A"), "B")

sim$run_up_to_size("B",1e4)
#>  [████████████████████████████████████████] 100% [00m:00s] Saving snapshot                                                                                                           

bbox <- sim$search_sample(c("A" = 100,"B" = 100), 50, 50)
sim$sample_cells("Sampling", bbox$lower_corner, bbox$upper_corner)
forest = sim$get_samples_forest()
labels = get_events_table(forest)

m_engine = build_mutation_engine(setup_code = "demo")
#> Downloading reference genome...
#> Reference genome downloaded
#> Decompressing reference file...done
#> Downloading SBS file...
#> SBS file downloaded
#> Downloading driver mutation file...
#> Driver mutation file downloaded
#> Downloading passenger CNAs file...
#> Passenger CNAs file downloaded
#> Downloading germline mutations...
#> Germline mutations downloaded
#> Building context index...
#>  [█---------------------------------------] 0% [00m:00s] Processing chr. 22                                                                                                           [█████████████████-----------------------] 40% [00m:01s] Processing chr. 22                                                                                                          [█████████████████████████████████-------] 81% [00m:02s] Processing chr. 22                                                                                                          [████████████████████████████████████████] 100% [00m:02s] Context index built                                                                                                       
#>  [█---------------------------------------] 0% [00m:00s] Saving context index                                                                                                         [████████████████████████████████████████] 100% [00m:00s] Context index saved                                                                                                       
#> done
#>  [█---------------------------------------] 0% [00m:00s] Loading germline                                                                                                             [████████████████████████████████████████] 100% [00m:00s] Germline loaded                                                                                                           
#>  [█---------------------------------------] 0% [00m:00s] Saving germline                                                                                                              [████████████████████████████████████████] 100% [00m:00s] Germline saved                                                                                                            
m_engine$add_mutant(mutant_name = "A",
                    passenger_rates = c(SNV = 5e-8))
#>  [█---------------------------------------] 0% [00m:00s] Retrieving "A" SNVs                                                                                                          [████████████████████████████████████████] 100% [00m:00s] "A" SNVs retrieved                                                                                                        
m_engine$add_mutant(mutant_name = "B",
                    passenger_rates = c(SNV = 5e-8))
#>  [█---------------------------------------] 0% [00m:00s] Retrieving "B" SNVs                                                                                                          [████████████████████████████████████████] 100% [00m:00s] "B" SNVs retrieved                                                                                                        

m_engine$add_exposure(time = 0, c(SBS1 = 0.2,SBS5 = 0.8))

phylo_forest = m_engine$place_mutations(forest, 100)
#>  [█---------------------------------------] 0% [00m:00s] Placing mutations                                                                                                            [████████████████████████████████████████] 100% [00m:00s] Mutations placed                                                                                                          

labels = get_events_table(forest)

seq_results = simulate_seq(phylo_forest, coverage = 5,
                           epi_FACS = F, write_SAM = F)
#>  [█---------------------------------------] 0% [00m:00s] Found 22                                                                                                                     [█---------------------------------------] 0% [00m:00s] Reading 22                                                                                                                   [█---------------------------------------] 0% [00m:01s] Reading 22                                                                                                                   [█---------------------------------------] 0% [00m:01s] Processing chr. 22                                                                                                           [██--------------------------------------] 3% [00m:02s] Processing chr. 22                                                                                                           [███████---------------------------------] 16% [00m:03s] Processing chr. 22                                                                                                          [████████████----------------------------] 28% [00m:04s] Processing chr. 22                                                                                                          [█████████████████-----------------------] 40% [00m:05s] Processing chr. 22                                                                                                          [██████████████████████------------------] 52% [00m:06s] Processing chr. 22                                                                                                          [███████████████████████████-------------] 65% [00m:07s] Processing chr. 22                                                                                                          [███████████████████████████████---------] 77% [00m:08s] Processing chr. 22                                                                                                          [█████████████████████████████████████---] 90% [00m:09s] Processing chr. 22                                                                                                          [████████████████████████████████████████] 99% [00m:10s] Processing chr. 22                                                                                                          [████████████████████████████████████████] 99% [00m:11s] Processing chr. 22                                                                                                          [████████████████████████████████████████] 99% [00m:12s] Processing chr. 22                                                                                                          [████████████████████████████████████████] 99% [00m:13s] Processing chr. 22                                                                                                          [████████████████████████████████████████] 99% [00m:14s] Processing chr. 22                                                                                                          [████████████████████████████████████████] 99% [00m:15s] Processing chr. 22                                                                                                          [████████████████████████████████████████] 99% [00m:16s] Processing chr. 22                                                                                                          [████████████████████████████████████████] 99% [00m:17s] Processing chr. 22                                                                                                          [████████████████████████████████████████] 100% [00m:18s] Read simulated                                                                                                            

muts_to_sticks(phylo_forest, seq_results, labels)
#> # A tibble: 916 × 20
#>    chr    chr_pos ref   alt   causes classes   Sampling.occurrences
#>    <chr>    <int> <chr> <chr> <chr>  <chr>                    <int>
#>  1 22    16066239 G     A     SBS1   passenger                    1
#>  2 22    16070145 G     T     SBS5   passenger                    1
#>  3 22    16076492 G     T     SBS5   passenger                   22
#>  4 22    16087282 G     A     SBS5   passenger                    1
#>  5 22    16088873 C     A     SBS5   passenger                    1
#>  6 22    16091480 C     T     SBS5   passenger                    2
#>  7 22    16095931 T     C     SBS5   passenger                    1
#>  8 22    16098576 G     A     SBS5   passenger                   16
#>  9 22    16100068 C     T     SBS1   passenger                    1
#> 10 22    16101939 A     G     SBS5   passenger                    1
#> # ℹ 906 more rows
#> # ℹ 13 more variables: Sampling.coverage <int>, Sampling.VAF <dbl>,
#> #   normal_sample.occurrences <int>, normal_sample.coverage <int>,
#> #   normal_sample.VAF <dbl>, id <chr>, cell_id <dbl>, ancestor <int>,
#> #   mutant <chr>, epistate <chr>, sample <chr>, birth_time <dbl>, obs <chr>