Add a mutant specification to the mutation engine.

This method adds a mutant specification to the mutation engine. The users must use it to specify the name and the genomic characterization (i.e., SNVs and CNAs) of all the simulated mutants together with the mutation rates of its species.

Arguments

mutant_name

The mutant name.

passenger_rates

The list of the passenger rates whose names are the epigenetic states of the species or a single rate, if the mutant does not have an epigenetic state.

driver_SNVs

The list of the driver SNVs characterizing the mutant.

driver_CNAs

The list of the driver CNAs characterizing the mutant.

Examples

# create a demostrative mutation engine
m_engine <- build_mutation_engine(setup_code = "demo")

# add the mutant "A" characterized by one driver SNV on chromosome 22 and
# two CNAs: an amplification and a deletion. The mutant has two epigenetic
# states and its species "A+" and "A-" have passenger SNV rates 1e-9 and
# 3e-8, respectively, and passenger CNA rates 0 and 1e-11, respectively.
m_engine$add_mutant("A", list("+" = c(SNV = 1e-9),
                              "-" = c(SNV = 3e-8, CNA = 1e-11)),
                    driver_SNVs = c(SNV("22", 10510210, "C")),
                    driver_CNAs = c(CNA(type = "A", "22", pos_in_chr = 10303470,
                                        len = 200000),
                                    CNA("D", "22", 5010000, 200000)))

# add the mutant "B" characterized by one driver SNV on chromosome 1 (no CNA)
# and missing epigenetic state. Its species "B" has passenger SNV rate 5e-9
# and passenger CNA rate 0.
m_engine$add_mutant("B", c(SNV = 5e-9), c(SNV("1", 10510210, "C")))
#> Error: SNV chr1(10510210)(?[?>C]?) belongs to an unknown chromosome

m_engine
#> MutationEngine
#>  Passenger rates
#>    "A+": {SNV: 1e-09}
#>    "A-": {SNV: 3e-08, CNA: 1e-11}
#> 
#>  Driver mutations
#>    "A" SNVs: 
#>        chr22(10510210)(A[T>C]T)
#>    "A" CNAs: 
#>        CNA("D",GenomicRegion(chr: 22, begin: 5010000, length: 200000))
#>        CNA("A",GenomicRegion(chr: 22, begin: 10303470, length: 200000))
#> 
#>  Timed Exposure
#>