Get a data frame containing the available SBSs

This method returns a data frame containing the available SBSs and the corresponding mutation probability. The first column ("Type") describes a mutation in a context, while each of the remaining columns contains the probabilities of the mutations for one of the available SBSs.

Value

A data frame containing the available SBSs.

Examples

# build a mutation engine
m_engine <- build_mutation_engine(setup_code = "demo")
#>  [█---------------------------------------] 0% [00m:00s] Loading context index                                   [████████████████████████████████████████] 100% [00m:00s] Context index loaded                                 
#>  [█---------------------------------------] 0% [00m:00s] Loading germline                                        [████████████████████████████████████████] 100% [00m:00s] Germline loaded                                      

# get the SBS data frame
head(m_engine$get_SBSs(), 5)
#>      Type        SBS1        SBS2       SBS3        SBS4        SBS5
#> 1 A[C>A]A 0.000886157 5.80000e-07 0.02080832 0.042196498 0.011997600
#> 2 A[C>A]C 0.002280405 1.48004e-04 0.01650660 0.033297236 0.009438112
#> 3 A[C>A]G 0.000177031 5.23000e-05 0.00175070 0.015598705 0.001849630
#> 4 A[C>A]T 0.001280227 9.78000e-05 0.01220488 0.029497552 0.006608678
#> 5 A[C>G]A 0.001860330 2.23000e-16 0.01970788 0.006889428 0.010097980
#>          SBS6       SBS7a       SBS7b       SBS7c       SBS7d        SBS8
#> 1 0.000425233 0.000067000 0.002329386 0.004830406 0.000040400 0.044098218
#> 2 0.000524287 0.000179116 0.000460879 0.001150097 0.000764883 0.047798069
#> 3 0.000052000 0.000071200 0.000185951 0.000377032 0.000249962 0.004619813
#> 4 0.000180099 0.000248161 0.000709813 0.001960165 0.004049378 0.046998101
#> 5 0.000471258 0.000064900 0.000008560 0.001120094 0.001179819 0.004329825
#>          SBS9       SBS10a      SBS10b      SBS10c      SBS10d       SBS11
#> 1 0.000557954 2.190170e-03 0.000181997 0.004331453 0.010113971 0.000146208
#> 2 0.004089661 1.770137e-03 0.006539908 0.014830220 0.018446062 0.000552786
#> 3 0.000425965 1.500120e-04 0.000053500 0.000657116 0.000726535 0.000094200
#> 4 0.003049747 1.700132e-02 0.000016300 0.013127763 0.014196655 0.000266379
#> 5 0.004839598 2.230000e-16 0.000304996 0.000347634 0.000128897 0.000170242
#>         SBS12       SBS13        SBS14       SBS15       SBS16      SBS17a
#> 1 0.004520592 0.001819920 1.120492e-03 0.000944202 0.015995037 0.002070261
#> 2 0.001130148 0.000720968 1.310575e-02 0.000497106 0.002899100 0.000918116
#> 3 0.000540071 0.000263988 4.131810e-04 0.000046100 0.001019684 0.000047600
#> 4 0.001220160 0.000347985 8.263628e-02 0.001110238 0.010596712 0.000061800
#> 5 0.001220160 0.003869830 2.220000e-16 0.000113024 0.001739460 0.001010128
#>        SBS17b       SBS18       SBS19       SBS20       SBS21      SBS22a
#> 1 0.000607958 0.051533859 0.001269382 0.000619283 0.000156993 0.006046172
#> 2 0.000128991 0.015810387 0.000640688 0.001390636 0.002359889 0.000094500
#> 3 0.000058200 0.002431598 0.000245880 0.000021800 0.000293986 0.000790362
#> 4 0.000455969 0.021414070 0.000570722 0.001240568 0.000620971 0.001729623
#> 5 0.000145990 0.001731137 0.003188447 0.008774017 0.000016900 0.001669991
#>        SBS22b       SBS23      SBS24       SBS25       SBS26       SBS27
#> 1 0.006086541 8.35523e-04 0.03641340 0.009898466 0.000873118 0.005207818
#> 2 0.003802506 3.99250e-04 0.02650976 0.006998915 0.000528071 0.004738015
#> 3 0.000980017 9.86000e-08 0.01480545 0.001449775 0.000114015 0.000782672
#> 4 0.003566008 5.61000e-18 0.02200810 0.004969230 0.000619084 0.002718861
#> 5 0.004179338 5.74000e-18 0.00334123 0.008028756 0.000427058 0.001319447
#>         SBS28      SBS29       SBS30       SBS31       SBS32       SBS33
#> 1 0.000783905 0.06320926 0.001799681 0.009534985 0.022297772 0.003111360
#> 2 0.002529694 0.05120751 0.000505910 0.018490274 0.018798122 0.002230975
#> 3 0.000352957 0.01920282 0.000091300 0.001659127 0.004459554 0.000414181
#> 4 0.003969520 0.03560522 0.000555901 0.006276698 0.014598542 0.001900831
#> 5 0.000153981 0.00314046 0.001069810 0.008315626 0.001989801 0.001410616
#>         SBS34       SBS35       SBS36       SBS37       SBS38       SBS39
#> 1 0.004867949 0.008826949 0.025193777 0.003950822 0.012794869 0.011701521
#> 2 0.006957070 0.046184032 0.008317945 0.001450302 0.010095952 0.007150930
#> 3 0.000052000 0.001389520 0.002239447 0.001060221 0.001899238 0.002670347
#> 4 0.001229482 0.021592534 0.017895580 0.001850385 0.008846453 0.007400962
#> 5 0.000761679 0.003578763 0.001839546 0.034307136 0.002688922 0.047306150
#>        SBS40a      SBS40b      SBS40c       SBS41       SBS42       SBS43
#> 1 0.036395305 0.014007124 0.003757760 0.002110193 0.001160252 0.029506295
#> 2 0.016772493 0.004719804 0.006283197 0.001220112 0.020604480 0.005901259
#> 3 0.003747968 0.001139655 0.002019124 0.000061400 0.000033400 0.000626134
#> 4 0.015434802 0.004952959 0.008175782 0.001330122 0.007971733 0.002070442
#> 5 0.008213432 0.007453156 0.008893380 0.005330488 0.000406088 0.000361077
#>          SBS44       SBS45  SBS46       SBS47       SBS48       SBS49
#> 1 7.680000e-18 0.009108136 0.0044 0.067787870 0.000855468 0.025116633
#> 2 1.500380e-04 0.002849417 0.0047 0.029694686 0.000841460 0.007034659
#> 3 9.160000e-07 0.001659660 0.0003 0.001379753 0.000258141 0.007114712
#> 4 5.781465e-03 0.009638027 0.0034 0.011897871 0.000038900 0.007975282
#> 5 3.180806e-03 0.003109364 0.0012 0.001029816 0.000198108 0.000288191
#>         SBS50       SBS51       SBS52       SBS53       SBS54       SBS55
#> 1 0.119000274 0.140963913 0.015196277 0.005383510 0.002160358 0.005881127
#> 2 0.127000292 0.001689567 0.006538398 0.001961279 0.000796132 0.002050393
#> 3 0.005770013 0.004798772 0.004138986 0.038124854 0.001640271 0.000044900
#> 4 0.083200191 0.021194574 0.009237737 0.000017100 0.000414069 0.001170224
#> 5 0.003850009 0.003958986 0.001729576 0.007034586 0.001550257 0.002850546
#>         SBS56       SBS57       SBS58       SBS59        SBS60       SBS84
#> 1 0.012597481 0.012306325 0.058874618 0.003588208 6.152708e-03 0.003471995
#> 2 0.015696861 0.001430735 0.006747091 0.002368817 7.793430e-04 0.005007309
#> 3 0.000205959 0.003331713 0.000824644 0.000141929 2.230000e-16 0.000452642
#> 4 0.022995401 0.008644443 0.005577595 0.014492762 4.441950e-04 0.009295947
#> 5 0.000417916 0.015407920 0.020591123 0.002938533 3.771660e-04 0.006758076
#>         SBS85       SBS86       SBS87        SBS88       SBS89       SBS90
#> 1 0.006080257 0.002954301 0.008972551 1.000000e-18 0.032168854 0.002201650
#> 2 0.000879937 0.003774501 0.004572575 1.000000e-18 0.017693728 0.000708223
#> 3 0.000306011 0.000385490 0.006208566 1.000000e-18 0.009670834 0.000138566
#> 4 0.002717279 0.003624001 0.004957473 1.731102e-03 0.020743819 0.001755161
#> 5 0.007234927 0.052516013 0.007866457 1.000000e-18 0.014817169 0.000508073
#>         SBS91      SBS92       SBS93       SBS94       SBS95       SBS96
#> 1 0.002944550 0.01132934 0.011573228 0.015579670 0.038408380 0.011430605
#> 2 0.052997084 0.00974473 0.008096344 0.024746217 0.017383837 0.007770735
#> 3 0.000204200 0.00469657 0.001760621 0.001574271 0.008360147 0.001997198
#> 4 0.000130810 0.00775798 0.008420819 0.011076145 0.023294307 0.011809258
#> 5 0.000243319 0.00305586 0.008856798 0.007003872 0.003616709 0.009294594
#>         SBS97       SBS98       SBS99
#> 1 0.008159180 0.008804153 0.000003820
#> 2 0.007512896 0.006768714 0.000006960
#> 3 0.001374676 0.032959350 0.000051400
#> 4 0.006858397 0.003758009 0.000030700
#> 5 0.002346944 0.006896898 0.000446027