Get a data frame containing the available indel signatures

This method returns a data frame containing the available indel signatures and the corresponding mutation probability. The first column ("Type") describes a mutation in a context, while each of the remaining columns contains the probabilities of the mutations for one of the available indel signatures.

Value

A data frame containing the available indel signatures.

Examples

# build a mutation engine
m_engine <- build_mutation_engine(setup_code = "demo")
#>  [█---------------------------------------] 0% [00m:00s] Loading context index                                                                                                                                       [████████████████████████████████████████] 100% [00m:00s] Context index loaded                                                                                                                                     
#>  [█---------------------------------------] 0% [00m:00s] Loading RS index                                                                                                                                            [████████████████████--------------------] 48% [00m:01s] Loading RS index                                                                                                                                           [█████████████████████████████████████---] 92% [00m:02s] Loading RS index                                                                                                                                           [████████████████████████████████████████] 100% [00m:02s] RS index loaded                                                                                                                                          
#>  [█---------------------------------------] 0% [00m:00s] Loading germline                                                                                                                                            [████████████████████████████████████████] 100% [00m:00s] Germline loaded                                                                                                                                          

# get the indel data frame
head(m_engine$get_indel_signatures(), 5)
#>        Type          ID1         ID2        ID3         ID4         ID5
#> 1 1:Del:C:0 1.598890e-04 0.004824116 0.12472711 0.007249717 0.022202108
#> 2 1:Del:C:1 7.735230e-04 0.000022100 0.20887617 0.002734869 0.028547215
#> 3 1:Del:C:2 3.310000e-18 0.000003110 0.17632422 0.002041063 0.026596927
#> 4 1:Del:C:3 1.907613e-03 0.002472076 0.06404276 0.001112283 0.014159122
#> 5 1:Del:C:4 7.059900e-04 0.003856976 0.04398981 0.001075684 0.002873422
#>           ID6          ID7         ID8         ID9        ID10         ID11
#> 1 0.030506799 4.665610e-04 0.039827558 0.334939880 0.025105414 1.576565e-03
#> 2 0.005491903 1.000000e-18 0.023915938 0.099994226 0.029306320 5.503180e-04
#> 3 0.008033994 8.561945e-03 0.003330554 0.002690655 0.018804056 3.573380e-04
#> 4 0.001102379 4.236015e-02 0.005403153 0.010625416 0.016503560 9.080000e-18
#> 5 0.002478170 1.148919e-01 0.001275709 0.012496902 0.002930632 9.080000e-18
#>          ID12        ID13        ID14       ID15         ID16         ID17
#> 1 0.052787741 0.065548371 0.012016009 0.02652737 9.527690e-04 0.0061175990
#> 2 0.014674665 0.017536463 0.001946232 0.12100979 2.640000e-18 0.0036675550
#> 3 0.000561932 0.005120219 0.000683696 0.14782886 8.150000e-05 0.0012203270
#> 4 0.000995577 0.005350516 0.000487468 0.13600606 1.624914e-03 0.0000000226
#> 5 0.002240050 0.004346005 0.002071921 0.08419770 3.378643e-03 0.0000000194
#>           ID18        ID19        ID20        ID21        ID22       ID23
#> 1 6.012048e-03 0.054089076 0.006578444 0.001388870 0.002569923 0.05169788
#> 2 2.004016e-03 0.015008861 0.003268685 0.004836623 0.001730933 0.10619410
#> 3 2.004016e-03 0.001849447 0.002094676 0.001284341 0.000798746 0.05287559
#> 4 4.008032e-03 0.006563183 0.002189543 0.002792894 0.001532915 0.03862287
#> 5 1.000000e-18 0.008206888 0.000481952 0.005253476 0.002421533 0.01125206