The phylogenetic forest of cells in samples.

Represents the phylogenetic forest of the cells sampled during the computation. The leaves of this forest are the sampled cells. This class is analoguous to the class `SamplesForest`, but each node is labelled with the mutations occuring for the first time on the cell represented by the node itself. Moreover each leaf is also associated with the genome mutations occurring in the corresponding cell.

Fields

get_coalescent_cells

Retrieve most recent common ancestors

Parameter: cell_ids - The list of the identifiers of the cells whose most recent common ancestors are aimed (optional).

Return: A data frame representing, for each of the identified cells, the identified (column "cell_id"), whenever the node is not a root, the ancestor identifier (column "ancestor"), whenever the node was sampled, i.e., it is one of the forest leaves, the name of the sample containing the node, (column "sample"), the mutant (column "mutant"), the epistate (column "epistate"), and the birth time (column "birth_time").

get_first_occurrences

Gets the identifier of the cell in which a mutation occurs for the first time

Parameter: mutation - A mutation being either a SNV or a CNA.

Return: The identifier of the cell in which a mutation occurs for the first time

get_nodes

Get the forest nodes

Return: A data frame representing, for each node in the forest, the identified (column "id"), whenever the node is not a root, the ancestor identifier (column "ancestor"), whenever the node was sampled, i.e., it is one of the forest leaves, the name of the sample containing the node, (column "sample"), the mutant (column "mutant"), the epistate (column "epistate"), and the birth time (column "birth_time").

get_sampled_cell_SNVs

Gets a the SNVs of the sampled cells

Returns: A data frame reporting `cell_id`, `chromosome`, `chr_pos` (i.e., position in the chromosome), `allele` (in which the SNV occurs), `context`, `mutated_base`, and `cause` for each SNV in the sampled cell genomes.

get_samples_info

Retrieve information about the samples

Returns: A data frame containing, for each sample collected during the simulation, the columns "name", "time", "ymin", "xmin", "ymax", "xmax", and "tumoral cells". "ymin", "xmin", "ymax", "xmax" report the boundaries of the sampled rectangular region, while "tumoral cells" is the number of tumoral cells in the sample.

get_species_info

Gets the species data

Returns: A data frame reporting "mutant" and "epistate" for each registered species.

get_subforest_for

Build a subforest using as leaves some of the original samples

Parameter: sample_names - The names of the samples whose cells will be used as leaves of the new forest.

Returns: A samples forest built on the samples mentioned in `sample_names`.

save

Save a phylogenetic forest in a file

Parameter: filename - The path of the file in which the phylogenetic forest must be saved.