Compute the forest sticks

A crucial node of a forest is a root of the forest, a node whose parent belongs to a different species, or the most recent common ancestor of two crucial nodes.

A stick is a path of the forest in which the only crucial nodes are the first and the last one.

This method return the list of the forest sticks. Each stick is represented by the sequence of cell identifiers labelling the nodes in the stick.

Arguments

birth_threshold

The maximum birth time for the cells associated to the returned sticks (optional).

Value

The list of the forest sticks whose associated cells have birth time smaller than or equal to birth_threshold. Each stick is represented as the list of cell identifiers labelling the nodes in the stick from the higher to the deeper in the forest.

See also

Examples

sim <- new(Simulation)
sim$add_mutant(name = "A", growth_rate = 0.2,
               death_rate = 0.01)
sim$place_cell("A", 500, 500)

sim$death_activation_level <- 100
sim$run_up_to_size(species = "A", num_of_cells = 5000)
#>  [████████████████████████████████████████] 100% [00m:00s] Saving snapshot                                                                                                                                    

sim$get_clock()
#> [1] 110.8153

sim$add_mutant(name = "B", growth_rate = 0.3, death_rate = 0.01)
sim$mutate_progeny(sim$choose_cell_in("A"), "B")
sim$run_up_to_size(species = "B", num_of_cells = 50)
#>  [████████████████████████████████████████] 100% [00m:00s] Saving snapshot                                                                                                                                    

sim$get_clock()
#> [1] 125.5882

sim$add_mutant(name = "C", growth_rate = 0.3, death_rate = 0.01)
sim$mutate_progeny(sim$choose_cell_in("B"), "C")
sim$run_up_to_size(species = "C", num_of_cells = 100)
#>  [████████████████████████████████████████] 100% [00m:00s] Saving snapshot                                                                                                                                    

# search for a 33x33 region containing 50 cells in A and
# 50 cells in B at least and sample it
region <- sim$search_sample(c(A = 50, B = 50), 33, 33)
sim$sample_cells("S1", region$lower_corner, region$upper_corner)

# search for a 33x33 region containing 50 cells in B and
# 50 cells in C at least and sample it
region <- sim$search_sample(c(B = 50, C = 50), 33, 33)
sim$sample_cells("S2", region$lower_corner, region$upper_corner)

# build the samples forest
forest <- sim$get_samples_forest()

# search for the forest sticks
forest$get_sticks()
#> [[1]]
#>  [1] 12517 12779 13321 14355 15818 16193 16224 17216 17372 21025
#> 
#> [[2]]
#>  [1]     0     1     3    10    20    24    42   104   165   170   194   312
#> [13]   315   342   367   444   566   570   798   833  1072  1135  1158  1214
#> [25]  1600  1664  2068  2573  2764  2938  2988  3260  3288  3684  3768  3838
#> [37]  4262  4440  4717  5057  5062  6043  9430  9536 10198 10770 10782 11014
#> [49] 12188 12517
#> 

# search for the forest sticks whose corresponding cells have 
# birth times 120.2 time units at most
forest$get_sticks(120.2)
#> [[1]]
#>  [1]     0     1     3    10    20    24    42   104   165   170   194   312
#> [13]   315   342   367   444   566   570   798   833  1072  1135  1158  1214
#> [25]  1600  1664  2068  2573  2764  2938  2988  3260  3288  3684  3768  3838
#> [37]  4262  4440  4717  5057  5062  6043  9430  9536 10198 10770 10782 11014
#> [49] 12188 12517
#>