Gets the tissue cells

This method collects some data about the cells in the tissue without altering the tissue itself. The pairs of optional parameters `lower_corner` and `upper_corner` define a frame of the tissue in which the data are sampled. The optional parameters `mutant_filter` and `epigenetic_filter` filter the collected cell data according to the cell mutant and epigenetic state.

Arguments

lower_corner

The lower-left corner of the selection frame (optional).

upper_corner

The upper-right corner of the selection frame (optional).

mutant_filter

The vector of the to-be-selected mutant names (optional).

epigenetic_filter

The vector of the to-be-selected epigenetic states (optional).

Value

A data frame reporting "cell_id", "mutant", "epistate", "position_x", and "position_y" for each cells satisfying the provided filters and laying in the input frame.

Examples

sim <- new(Simulation)
sim$add_mutant(name = "A",
                 epigenetic_rates = c("+-" = 0.01, "-+" = 0.01),
                 growth_rates = c("+" = 0.2, "-" = 0.08),
                 death_rates = c("+" = 0.1, "-" = 0.01))
sim$add_mutant(name = "B",
                 epigenetic_rates = c("+-" = 0.02, "-+" = 0.01),
                 growth_rates = c("+" = 0.3, "-" = 0.1),
                 death_rates = c("+" = 0.1, "-" = 0.01))
sim$schedule_mutation(src = "A", dst = "B", time = 50)
sim$place_cell("A+", 500, 500)
sim$run_up_to_time(30)

# collect all the cells in the tissue
sim$get_cells()
#>    cell_id mutant epistate position_x position_y
#> 1      240      A        +        492        501
#> 2      252      A        -        493        500
#> 3      239      A        +        493        501
#> 4      179      A        +        493        503
#> 5      251      A        -        494        501
#> 6      192      A        +        494        502
#> 7      214      A        +        495        499
#> 8      191      A        +        495        500
#> 9      244      A        +        495        501
#> 10     180      A        +        495        502
#> 11     158      A        +        495        503
#> 12     222      A        +        495        505
#> 13     187      A        +        496        494
#> 14     218      A        +        496        495
#> 15     175      A        +        496        499
#> 16     213      A        -        496        500
#> 17     243      A        +        496        501
#> 18     198      A        +        496        502
#> 19     221      A        +        496        503
#> 20     228      A        +        496        504
#> 21     188      A        +        497        494
#> 22     217      A        +        497        496
#> 23     166      A        +        497        497
#> 24     176      A        +        497        498
#> 25     212      A        +        497        499
#> 26     197      A        +        497        500
#> 27     223      A        +        497        501
#> 28     224      A        +        497        502
#> 29     236      A        +        497        503
#> 30     195      A        +        498        497
#> 31     211      A        +        498        498
#> 32     207      A        +        498        499
#> 33     201      A        +        498        500
#> 34     246      A        +        498        501
#> 35     235      A        +        498        502
#> 36     227      A        +        498        503
#> 37     210      A        +        499        496
#> 38     196      A        +        499        497
#> 39     204      A        +        499        498
#> 40     208      A        +        499        499
#> 41     181      A        +        499        500
#> 42     245      A        +        499        501
#> 43     186      A        +        499        503
#> 44     200      A        +        500        495
#> 45     209      A        +        500        497
#> 46     203      A        +        500        499
#> 47     182      A        +        500        500
#> 48      81      A        -        500        501
#> 49     247      A        +        500        502
#> 50     248      A        +        500        503
#> 51     199      A        +        501        496
#> 52     169      A        +        501        499
#> 53     168      A        +        501        500
#> 54     161      A        -        501        501
#> 55      82      A        -        501        502
#> 56     184      A        +        501        503
#> 57     219      A        +        502        499
#> 58     205      A        +        502        500
#> 59     189      A        +        502        501
#> 60     190      A        +        502        502
#> 61     220      A        +        503        498
#> 62     162      A        -        503        499
#> 63     254      A        +        503        500
#> 64     206      A        +        503        501
#> 65     225      A        +        503        502
#> 66     238      A        +        504        495
#> 67     233      A        +        504        497
#> 68     229      A        +        504        498
#> 69     230      A        +        504        499
#> 70     253      A        +        504        500
#> 71     127      A        -        504        501
#> 72     241      A        +        504        502
#> 73     226      A        +        504        503
#> 74     237      A        +        505        496
#> 75     234      A        +        505        497
#> 76     151      A        +        505        499
#> 77     171      A        +        505        500
#> 78     242      A        +        505        501
#> 79     174      A        +        506        497
#> 80     250      A        -        506        498
#> 81     249      A        +        506        499
#> 82     215      A        +        506        501
#> 83     216      A        +        506        502
#> 84     232      A        +        507        497
#> 85     177      A        +        507        501

# get the cells in the frame [495,505]x[490,500]
sim$get_cells(lower_corner=c(495,490), upper_corner=c(505,500))
#>    cell_id mutant epistate position_x position_y
#> 1      214      A        +        495        499
#> 2      191      A        +        495        500
#> 3      187      A        +        496        494
#> 4      218      A        +        496        495
#> 5      175      A        +        496        499
#> 6      213      A        -        496        500
#> 7      188      A        +        497        494
#> 8      217      A        +        497        496
#> 9      166      A        +        497        497
#> 10     176      A        +        497        498
#> 11     212      A        +        497        499
#> 12     197      A        +        497        500
#> 13     195      A        +        498        497
#> 14     211      A        +        498        498
#> 15     207      A        +        498        499
#> 16     201      A        +        498        500
#> 17     210      A        +        499        496
#> 18     196      A        +        499        497
#> 19     204      A        +        499        498
#> 20     208      A        +        499        499
#> 21     181      A        +        499        500
#> 22     200      A        +        500        495
#> 23     209      A        +        500        497
#> 24     203      A        +        500        499
#> 25     182      A        +        500        500
#> 26     199      A        +        501        496
#> 27     169      A        +        501        499
#> 28     168      A        +        501        500
#> 29     219      A        +        502        499
#> 30     205      A        +        502        500
#> 31     220      A        +        503        498
#> 32     162      A        -        503        499
#> 33     254      A        +        503        500
#> 34     238      A        +        504        495
#> 35     233      A        +        504        497
#> 36     229      A        +        504        498
#> 37     230      A        +        504        499
#> 38     253      A        +        504        500
#> 39     237      A        +        505        496
#> 40     234      A        +        505        497
#> 41     151      A        +        505        499
#> 42     171      A        +        505        500

# cells can be filtered by mutant name...
sim$get_cells(mutant_filter=c("A"),epigenetic_filter=c("+","-"))
#>    cell_id mutant epistate position_x position_y
#> 1      240      A        +        492        501
#> 2      252      A        -        493        500
#> 3      239      A        +        493        501
#> 4      179      A        +        493        503
#> 5      251      A        -        494        501
#> 6      192      A        +        494        502
#> 7      214      A        +        495        499
#> 8      191      A        +        495        500
#> 9      244      A        +        495        501
#> 10     180      A        +        495        502
#> 11     158      A        +        495        503
#> 12     222      A        +        495        505
#> 13     187      A        +        496        494
#> 14     218      A        +        496        495
#> 15     175      A        +        496        499
#> 16     213      A        -        496        500
#> 17     243      A        +        496        501
#> 18     198      A        +        496        502
#> 19     221      A        +        496        503
#> 20     228      A        +        496        504
#> 21     188      A        +        497        494
#> 22     217      A        +        497        496
#> 23     166      A        +        497        497
#> 24     176      A        +        497        498
#> 25     212      A        +        497        499
#> 26     197      A        +        497        500
#> 27     223      A        +        497        501
#> 28     224      A        +        497        502
#> 29     236      A        +        497        503
#> 30     195      A        +        498        497
#> 31     211      A        +        498        498
#> 32     207      A        +        498        499
#> 33     201      A        +        498        500
#> 34     246      A        +        498        501
#> 35     235      A        +        498        502
#> 36     227      A        +        498        503
#> 37     210      A        +        499        496
#> 38     196      A        +        499        497
#> 39     204      A        +        499        498
#> 40     208      A        +        499        499
#> 41     181      A        +        499        500
#> 42     245      A        +        499        501
#> 43     186      A        +        499        503
#> 44     200      A        +        500        495
#> 45     209      A        +        500        497
#> 46     203      A        +        500        499
#> 47     182      A        +        500        500
#> 48      81      A        -        500        501
#> 49     247      A        +        500        502
#> 50     248      A        +        500        503
#> 51     199      A        +        501        496
#> 52     169      A        +        501        499
#> 53     168      A        +        501        500
#> 54     161      A        -        501        501
#> 55      82      A        -        501        502
#> 56     184      A        +        501        503
#> 57     219      A        +        502        499
#> 58     205      A        +        502        500
#> 59     189      A        +        502        501
#> 60     190      A        +        502        502
#> 61     220      A        +        503        498
#> 62     162      A        -        503        499
#> 63     254      A        +        503        500
#> 64     206      A        +        503        501
#> 65     225      A        +        503        502
#> 66     238      A        +        504        495
#> 67     233      A        +        504        497
#> 68     229      A        +        504        498
#> 69     230      A        +        504        499
#> 70     253      A        +        504        500
#> 71     127      A        -        504        501
#> 72     241      A        +        504        502
#> 73     226      A        +        504        503
#> 74     237      A        +        505        496
#> 75     234      A        +        505        497
#> 76     151      A        +        505        499
#> 77     171      A        +        505        500
#> 78     242      A        +        505        501
#> 79     174      A        +        506        497
#> 80     250      A        -        506        498
#> 81     249      A        +        506        499
#> 82     215      A        +        506        501
#> 83     216      A        +        506        502
#> 84     232      A        +        507        497
#> 85     177      A        +        507        501

# ...or by epigenetic state
sim$get_cells(mutant_filter=c("A","B"),epigenetic_filter=c("-"))
#>   cell_id mutant epistate position_x position_y
#> 1     252      A        -        493        500
#> 2     251      A        -        494        501
#> 3     213      A        -        496        500
#> 4      81      A        -        500        501
#> 5     161      A        -        501        501
#> 6      82      A        -        501        502
#> 7     162      A        -        503        499
#> 8     127      A        -        504        501
#> 9     250      A        -        506        498

# cells can be filtered by frame, mutant, and epigenetic states
sim$get_cells(lower_corner=c(495,495), upper_corner=c(505,505),
              mutant_filter=c("A"),epigenetic_filter=c("+","-"))
#>    cell_id mutant epistate position_x position_y
#> 1      214      A        +        495        499
#> 2      191      A        +        495        500
#> 3      244      A        +        495        501
#> 4      180      A        +        495        502
#> 5      158      A        +        495        503
#> 6      222      A        +        495        505
#> 7      218      A        +        496        495
#> 8      175      A        +        496        499
#> 9      213      A        -        496        500
#> 10     243      A        +        496        501
#> 11     198      A        +        496        502
#> 12     221      A        +        496        503
#> 13     228      A        +        496        504
#> 14     217      A        +        497        496
#> 15     166      A        +        497        497
#> 16     176      A        +        497        498
#> 17     212      A        +        497        499
#> 18     197      A        +        497        500
#> 19     223      A        +        497        501
#> 20     224      A        +        497        502
#> 21     236      A        +        497        503
#> 22     195      A        +        498        497
#> 23     211      A        +        498        498
#> 24     207      A        +        498        499
#> 25     201      A        +        498        500
#> 26     246      A        +        498        501
#> 27     235      A        +        498        502
#> 28     227      A        +        498        503
#> 29     210      A        +        499        496
#> 30     196      A        +        499        497
#> 31     204      A        +        499        498
#> 32     208      A        +        499        499
#> 33     181      A        +        499        500
#> 34     245      A        +        499        501
#> 35     186      A        +        499        503
#> 36     200      A        +        500        495
#> 37     209      A        +        500        497
#> 38     203      A        +        500        499
#> 39     182      A        +        500        500
#> 40      81      A        -        500        501
#> 41     247      A        +        500        502
#> 42     248      A        +        500        503
#> 43     199      A        +        501        496
#> 44     169      A        +        501        499
#> 45     168      A        +        501        500
#> 46     161      A        -        501        501
#> 47      82      A        -        501        502
#> 48     184      A        +        501        503
#> 49     219      A        +        502        499
#> 50     205      A        +        502        500
#> 51     189      A        +        502        501
#> 52     190      A        +        502        502
#> 53     220      A        +        503        498
#> 54     162      A        -        503        499
#> 55     254      A        +        503        500
#> 56     206      A        +        503        501
#> 57     225      A        +        503        502
#> 58     238      A        +        504        495
#> 59     233      A        +        504        497
#> 60     229      A        +        504        498
#> 61     230      A        +        504        499
#> 62     253      A        +        504        500
#> 63     127      A        -        504        501
#> 64     241      A        +        504        502
#> 65     226      A        +        504        503
#> 66     237      A        +        505        496
#> 67     234      A        +        505        497
#> 68     151      A        +        505        499
#> 69     171      A        +        505        500
#> 70     242      A        +        505        501