Gets the tissue cells

This method collects some data about the cells in the tissue without altering the tissue itself. The pairs of optional parameters lower_corner and upper_corner define a frame of the tissue in which the data are sampled. The optional parameters mutant_filter and epigenetic_filter filter the collected cell data according to the cell mutant and epigenetic state.

Arguments

lower_corner

The lower-left corner of the selection frame (optional).

upper_corner

The upper-right corner of the selection frame (optional).

mutant_filter

The vector of the to-be-selected mutant names (optional).

epigenetic_filter

The vector of the to-be-selected epigenetic states (optional).

Value

A data frame reporting cell_id, mutant, epistate, position_x, and position_y for each cells satisfying the provided filters and laying in the input frame.

Examples

sim <- new(Simulation)
sim$add_mutant(name = "A",
               epigenetic_rates = c("+-" = 0.01, "-+" = 0.01),
               growth_rates = c("+" = 0.2, "-" = 0.08),
               death_rates = c("+" = 0.1, "-" = 0.01))
sim$add_mutant(name = "B",
               epigenetic_rates = c("+-" = 0.02, "-+" = 0.01),
               growth_rates = c("+" = 0.3, "-" = 0.1),
               death_rates = c("+" = 0.1, "-" = 0.01))
sim$schedule_mutation(src = "A", dst = "B", time = 50)
sim$place_cell("A+", 500, 500)
sim$run_up_to_time(30)
#>  [████████████████████████████████████████] 100% [00m:00s] Saving snapshot                                                                                                                

# collect all the cells in the tissue
sim$get_cells()
#>    cell_id mutant epistate position_x position_y
#> 1      346      A        +        496        507
#> 2      210      A        +        497        503
#> 3      251      A        +        497        505
#> 4      345      A        +        497        507
#> 5      336      A        +        498        500
#> 6      335      A        +        498        501
#> 7      344      A        +        498        502
#> 8      343      A        +        498        503
#> 9      310      A        +        498        504
#> 10     187      A        -        498        505
#> 11     289      A        +        498        506
#> 12     272      A        -        498        509
#> 13     242      A        +        499        496
#> 14      12      A        -        499        499
#> 15     348      A        -        499        500
#> 16     304      A        -        499        502
#> 17     225      A        +        499        503
#> 18     333      A        -        499        504
#> 19     236      A        +        499        505
#> 20     156      A        +        499        507
#> 21     271      A        -        499        508
#> 22     320      A        +        499        510
#> 23     275      A        +        500        497
#> 24     240      A        +        500        499
#> 25     347      A        -        500        500
#> 26     245      A        -        500        501
#> 27     296      A        +        500        502
#> 28     235      A        +        500        504
#> 29     290      A        +        500        505
#> 30     259      A        -        500        507
#> 31     323      A        -        500        509
#> 32     322      A        +        500        510
#> 33     319      A        +        500        511
#> 34     217      A        +        501        496
#> 35     247      A        +        501        498
#> 36     300      A        +        501        500
#> 37     222      A        +        501        502
#> 38     316      A        +        501        503
#> 39     107      A        -        501        505
#> 40     260      A        -        501        509
#> 41     324      A        -        501        510
#> 42     321      A        +        501        511
#> 43     288      A        +        502        497
#> 44     267      A        +        502        498
#> 45     330      A        +        502        499
#> 46     278      A        +        502        500
#> 47     338      A        +        502        501
#> 48     326      A        +        502        502
#> 49     192      A        +        502        503
#> 50     219      A        +        502        505
#> 51     317      A        +        502        506
#> 52     270      A        +        502        508
#> 53     284      A        +        502        509
#> 54     329      A        -        503        498
#> 55     273      A        +        503        499
#> 56     325      A        +        503        501
#> 57     337      A        +        503        502
#> 58     139      A        +        503        503
#> 59     266      A        +        503        504
#> 60     318      A        +        503        505
#> 61     195      A        -        503        506
#> 62     281      A        +        503        507
#> 63     283      A        +        503        508
#> 64     314      A        +        503        509
#> 65     298      A        +        503        512
#> 66     331      A        +        504        498
#> 67     237      A        -        504        499
#> 68     227      A        +        504        500
#> 69     233      A        +        504        501
#> 70     306      A        +        504        503
#> 71     265      A        +        504        504
#> 72     180      A        +        504        506
#> 73     262      A        +        504        509
#> 74     231      A        -        504        513
#> 75     212      A        +        505        498
#> 76     341      A        +        505        499
#> 77     234      A        +        505        500
#> 78     258      A        +        505        501
#> 79     190      A        +        505        503
#> 80     130      A        +        505        504
#> 81     280      A        +        505        505
#> 82     307      A        +        505        506
#> 83     342      A        +        506        498
#> 84     312      A        +        506        500
#> 85     311      A        +        506        501
#> 86     249      A        +        506        502
#> 87     308      A        +        506        506
#> 88     328      A        +        507        503
#> 89     286      A        +        507        506

# get the cells in the frame [495,505]x[490,500]
sim$get_cells(lower_corner=c(495,490), upper_corner=c(505,500))
#>    cell_id mutant epistate position_x position_y
#> 1      336      A        +        498        500
#> 2      242      A        +        499        496
#> 3       12      A        -        499        499
#> 4      348      A        -        499        500
#> 5      275      A        +        500        497
#> 6      240      A        +        500        499
#> 7      347      A        -        500        500
#> 8      217      A        +        501        496
#> 9      247      A        +        501        498
#> 10     300      A        +        501        500
#> 11     288      A        +        502        497
#> 12     267      A        +        502        498
#> 13     330      A        +        502        499
#> 14     278      A        +        502        500
#> 15     329      A        -        503        498
#> 16     273      A        +        503        499
#> 17     331      A        +        504        498
#> 18     237      A        -        504        499
#> 19     227      A        +        504        500
#> 20     212      A        +        505        498
#> 21     341      A        +        505        499
#> 22     234      A        +        505        500

# cells can be filtered by mutant name...
sim$get_cells(mutant_filter=c("A"),epigenetic_filter=c("+","-"))
#>    cell_id mutant epistate position_x position_y
#> 1      346      A        +        496        507
#> 2      210      A        +        497        503
#> 3      251      A        +        497        505
#> 4      345      A        +        497        507
#> 5      336      A        +        498        500
#> 6      335      A        +        498        501
#> 7      344      A        +        498        502
#> 8      343      A        +        498        503
#> 9      310      A        +        498        504
#> 10     187      A        -        498        505
#> 11     289      A        +        498        506
#> 12     272      A        -        498        509
#> 13     242      A        +        499        496
#> 14      12      A        -        499        499
#> 15     348      A        -        499        500
#> 16     304      A        -        499        502
#> 17     225      A        +        499        503
#> 18     333      A        -        499        504
#> 19     236      A        +        499        505
#> 20     156      A        +        499        507
#> 21     271      A        -        499        508
#> 22     320      A        +        499        510
#> 23     275      A        +        500        497
#> 24     240      A        +        500        499
#> 25     347      A        -        500        500
#> 26     245      A        -        500        501
#> 27     296      A        +        500        502
#> 28     235      A        +        500        504
#> 29     290      A        +        500        505
#> 30     259      A        -        500        507
#> 31     323      A        -        500        509
#> 32     322      A        +        500        510
#> 33     319      A        +        500        511
#> 34     217      A        +        501        496
#> 35     247      A        +        501        498
#> 36     300      A        +        501        500
#> 37     222      A        +        501        502
#> 38     316      A        +        501        503
#> 39     107      A        -        501        505
#> 40     260      A        -        501        509
#> 41     324      A        -        501        510
#> 42     321      A        +        501        511
#> 43     288      A        +        502        497
#> 44     267      A        +        502        498
#> 45     330      A        +        502        499
#> 46     278      A        +        502        500
#> 47     338      A        +        502        501
#> 48     326      A        +        502        502
#> 49     192      A        +        502        503
#> 50     219      A        +        502        505
#> 51     317      A        +        502        506
#> 52     270      A        +        502        508
#> 53     284      A        +        502        509
#> 54     329      A        -        503        498
#> 55     273      A        +        503        499
#> 56     325      A        +        503        501
#> 57     337      A        +        503        502
#> 58     139      A        +        503        503
#> 59     266      A        +        503        504
#> 60     318      A        +        503        505
#> 61     195      A        -        503        506
#> 62     281      A        +        503        507
#> 63     283      A        +        503        508
#> 64     314      A        +        503        509
#> 65     298      A        +        503        512
#> 66     331      A        +        504        498
#> 67     237      A        -        504        499
#> 68     227      A        +        504        500
#> 69     233      A        +        504        501
#> 70     306      A        +        504        503
#> 71     265      A        +        504        504
#> 72     180      A        +        504        506
#> 73     262      A        +        504        509
#> 74     231      A        -        504        513
#> 75     212      A        +        505        498
#> 76     341      A        +        505        499
#> 77     234      A        +        505        500
#> 78     258      A        +        505        501
#> 79     190      A        +        505        503
#> 80     130      A        +        505        504
#> 81     280      A        +        505        505
#> 82     307      A        +        505        506
#> 83     342      A        +        506        498
#> 84     312      A        +        506        500
#> 85     311      A        +        506        501
#> 86     249      A        +        506        502
#> 87     308      A        +        506        506
#> 88     328      A        +        507        503
#> 89     286      A        +        507        506

# ...or by epigenetic state
sim$get_cells(mutant_filter=c("A","B"),epigenetic_filter=c("-"))
#>    cell_id mutant epistate position_x position_y
#> 1      187      A        -        498        505
#> 2      272      A        -        498        509
#> 3       12      A        -        499        499
#> 4      348      A        -        499        500
#> 5      304      A        -        499        502
#> 6      333      A        -        499        504
#> 7      271      A        -        499        508
#> 8      347      A        -        500        500
#> 9      245      A        -        500        501
#> 10     259      A        -        500        507
#> 11     323      A        -        500        509
#> 12     107      A        -        501        505
#> 13     260      A        -        501        509
#> 14     324      A        -        501        510
#> 15     329      A        -        503        498
#> 16     195      A        -        503        506
#> 17     237      A        -        504        499
#> 18     231      A        -        504        513

# cells can be filtered by frame, mutant, and epigenetic states
sim$get_cells(lower_corner=c(495,495), upper_corner=c(505,505),
              mutant_filter=c("A"),epigenetic_filter=c("+","-"))
#>    cell_id mutant epistate position_x position_y
#> 1      210      A        +        497        503
#> 2      251      A        +        497        505
#> 3      336      A        +        498        500
#> 4      335      A        +        498        501
#> 5      344      A        +        498        502
#> 6      343      A        +        498        503
#> 7      310      A        +        498        504
#> 8      187      A        -        498        505
#> 9      242      A        +        499        496
#> 10      12      A        -        499        499
#> 11     348      A        -        499        500
#> 12     304      A        -        499        502
#> 13     225      A        +        499        503
#> 14     333      A        -        499        504
#> 15     236      A        +        499        505
#> 16     275      A        +        500        497
#> 17     240      A        +        500        499
#> 18     347      A        -        500        500
#> 19     245      A        -        500        501
#> 20     296      A        +        500        502
#> 21     235      A        +        500        504
#> 22     290      A        +        500        505
#> 23     217      A        +        501        496
#> 24     247      A        +        501        498
#> 25     300      A        +        501        500
#> 26     222      A        +        501        502
#> 27     316      A        +        501        503
#> 28     107      A        -        501        505
#> 29     288      A        +        502        497
#> 30     267      A        +        502        498
#> 31     330      A        +        502        499
#> 32     278      A        +        502        500
#> 33     338      A        +        502        501
#> 34     326      A        +        502        502
#> 35     192      A        +        502        503
#> 36     219      A        +        502        505
#> 37     329      A        -        503        498
#> 38     273      A        +        503        499
#> 39     325      A        +        503        501
#> 40     337      A        +        503        502
#> 41     139      A        +        503        503
#> 42     266      A        +        503        504
#> 43     318      A        +        503        505
#> 44     331      A        +        504        498
#> 45     237      A        -        504        499
#> 46     227      A        +        504        500
#> 47     233      A        +        504        501
#> 48     306      A        +        504        503
#> 49     265      A        +        504        504
#> 50     212      A        +        505        498
#> 51     341      A        +        505        499
#> 52     234      A        +        505        500
#> 53     258      A        +        505        501
#> 54     190      A        +        505        503
#> 55     130      A        +        505        504
#> 56     280      A        +        505        505